+++ standalone/src/test/scala/io/delta/standalone/internal/OptimisticTransactionLegacySuite.scala
@@ -33,7 +33,6 @@ import io.delta.standalone.types._
 
 import io.delta.standalone.internal.actions._
 import io.delta.standalone.internal.exception.DeltaErrors
-import io.delta.standalone.internal.sources.StandaloneHadoopConf
 import io.delta.standalone.internal.util.{ConversionUtils, SchemaUtils}
 import io.delta.standalone.internal.util.TestUtils._
 
@@ -257,32 +256,6 @@ class OptimisticTransactionLegacySuite extends FunSuite {
         }
         assert(e.getMessage.contains("Invalid Protocol"))
       }
-    }
-  }
-
-  test("Can't create table with external files") {
-    val extFile = AddFile("s3://snip/snip.parquet", Map.empty, 1, 1, dataChange = true)
-    val conf = new Configuration()
-    withTempDir { dir =>
-      val log = DeltaLog.forTable(conf, dir.getCanonicalPath)
-      val txn = log.startTransaction()
-      val e = intercept[IllegalStateException] {
-        txn.updateMetadata(metadataJ)
-        txn.commit(List(extFile), manualUpdate, engineInfo)
-      }
-      assert(e.getMessage.contains("Failed to relativize the path"))
-    }
-  }
-
-  test("Create table with external files override") {
-    val extFile = AddFile("s3://snip/snip.parquet", Map.empty, 1, 1, dataChange = true)
-    val conf = new Configuration()
-    conf.setBoolean(StandaloneHadoopConf.RELATIVE_PATH_IGNORE, true)
-    withTempDir { dir =>
-      val log = DeltaLog.forTable(conf, dir.getCanonicalPath)
-      val txn = log.startTransaction()
-      txn.updateMetadata(metadataJ)
-      txn.commit(List(extFile), manualUpdate, engineInfo)
     }
   }
